Refactor vfprintf, add vsnprintf
